requirements.txt
Django==1.9.4
Jinja2==2.8
MarkupSafe==0.23
WebOb==1.6.0
WebTest==2.0.20
apturl==0.5.2
arrow==0.7.0
beautifulsoup4==4.3.2
binaryornot==0.4.0
chardet==2.3.0
click==6.3
command-not-found==0.3
decorator==4.0.6
defer==1.0.6
django-bootstrap3==7.0.1
django-vanilla-views==1.0.4
django-webtest==1.7.8
future==0.15.2
html5lib==0.999
jinja2-time==0.1.0
language-selector==0.1
lxml==3.4.4
model-mommy==1.2.6
path.py==8.1.2
pbr==1.8.1
pexpect==3.2
pickleshare==0.5
poyo==0.1.0
pycups==1.9.73
pycurl==7.19.5.1
pygobject==3.16.2
pysmbc==1.0.15.4
python-dateutil==2.5.1
pyxdg==0.25
reportlab==3.2.0
simplegeneric==0.8.1
six==1.9.0
ufw==0.34
whichcraft==0.1.1
xkit==0.0.0
